Accept @font-face weight, width and oblique ranges and font-stretch percentages (#1445)
* Accept @font-face weight, width and oblique ranges and font-stretch percentages

A variable face covers its declared range in nearest-face matching, the weight, width and slant of the text set its wght, wdth and slnt axes inside the range, and nothing is faked that an axis supplies. A variable font with no descriptors covers the range of its own axes. The width is now a percentage end to end.

* Prefer an upright face over an oblique range for upright text and finish synthesis from the axes

Review fixes: an oblique-range face no longer beats a face that is upright itself, a variable face still fakes the bold or italic its axes cannot reach, and descriptor ranges split on any whitespace.

+++ src/PeachPDF/Html/Core/Utils/FontFamilyResolver.cs
@@ -12,7 +12,7 @@ namespace PeachPDF.Html.Core.Utils
     /// </summary>
     internal static class FontFamilyResolver
     {
-        internal static RFont? Resolve(RAdapter adapter, string fontFamilyList, double fsize, RFontStyle style, int? weight = null, int? stretch = null, double? obliqueSkewSinus = null, string? variations = null)
+        internal static RFont? Resolve(RAdapter adapter, string fontFamilyList, double fsize, RFontStyle style, int? weight = null, double? stretch = null, double? obliqueSkewSinus = null, string? variations = null)
         {
             var families = fontFamilyList.Split(',');
 
@@ -51,7 +51,7 @@ namespace PeachPDF.Html.Core.Utils
         /// <paramref name="presentation"/> is the emoji/text presentation the character was asked to be
         /// drawn in - see <see cref="Emoji.Resolve"/>.
         /// </summary>
-        internal static RFont? Resolve(RAdapter adapter, string fontFamilyList, double fsize, RFontStyle style, System.Text.Rune codepoint, int? weight = null, int? stretch = null, double? obliqueSkewSinus = null, PeachDrawing.Text.Unicode.EmojiPresentation presentation = PeachDrawing.Text.Unicode.EmojiPresentation.NoPreference, string? variations = null)
+        internal static RFont? Resolve(RAdapter adapter, string fontFamilyList, double fsize, RFontStyle style, System.Text.Rune codepoint, int? weight = null, double? stretch = null, double? obliqueSkewSinus = null, PeachDrawing.Text.Unicode.EmojiPresentation presentation = PeachDrawing.Text.Unicode.EmojiPresentation.NoPreference, string? variations = null)
         {
             // With a presentation request (CSS font-variant-emoji, or an explicit U+FE0E/U+FE0F), CSS Fonts 4
             // §5.3's cluster matching order applies: a family whose font supports the requested sequence
